% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbwadjust-package.R
\docType{package}
\name{sbwadjust-package}
\alias{sbwadjust}
\alias{sbwadjust-package}
\title{sbwadjust: Stable Balancing Weights for Covariate Adjustment in Trials}
\description{
Shared functions for the simulation studies and application in "Simple Covariate Adjustment for Many Estimands Using Stable Balancing Weights": fitting stable balancing weights (SBW), computing bootstrap-CI Kaplan-Meier survival-ratio estimates, and the covariate transforms and NHANES resampling helpers used to construct the simulation DGPs.
}
\author{
\strong{Maintainer}: Kayla Irish \email{kayla.a.irish@gmail.com}

}
\keyword{internal}
